chore(schema): rename pkg
