+++ apps/backend/package.json
@@ -16,7 +16,7 @@
     "@skylar/auth": "^0.1.0",
     "@skylar/db": "^0.1.0",
     "@skylar/logger": "^0.1.0",
-    "@skylar/schema": "^0.1.0",
+    "@skylar/parsers-and-types": "^0.1.0",
     "hono": "^3.5.6"
   },
   "devDependencies": {
