+++ apps/nextjs/package.json
@@ -24,7 +24,7 @@
     "@skylar/auth": "^0.1.0",
     "@skylar/logger": "^0.1.0",
     "@skylar/logic": "^0.1.0",
-    "@skylar/schema": "^0.1.0",
+    "@skylar/parsers-and-types": "^0.1.0",
     "@stripe/react-stripe-js": "^2.3.0",
     "@stripe/stripe-js": "^2.1.5",
     "@tanstack/react-query": "^4.33.0",
