+++ packages/api/package.json
@@ -14,7 +14,7 @@
   "dependencies": {
     "@skylar/auth": "^0.1.0",
     "@skylar/db": "^0.1.0",
-    "@skylar/schema": "^0.1.0",
+    "@skylar/parsers-and-types": "^0.1.0",
     "@trpc/client": "^10.38.1",
     "@trpc/server": "^10.38.1",
     "@tsndr/cloudflare-worker-jwt": "^2.2.2",
