+++ packages/auth/src/types/session.ts
@@ -1,4 +1,4 @@
-import type { UserType } from "@skylar/schema";
+import type { UserType } from "@skylar/parsers-and-types";
 
 export type Session = {
   user?: UserType;
