+++ packages/db/package.json
@@ -28,7 +28,7 @@
   },
   "dependencies": {
     "@neondatabase/serverless": "^0.6.0",
-    "@skylar/schema": "^0.1.0",
+    "@skylar/parsers-and-types": "^0.1.0",
     "drizzle-orm": "^0.28.5"
   },
   "devDependencies": {
